Add files via upload
This is the Arduino code for the BLE board to run the MiniMouse build. Drive pins are D5 and D6, Steering is D9, this utilised "software serial" configuration for the BLE board on pins D10 and D11 (Jumper position "B"). Note we use the "Dabble" and the DFRobotDFPlayerMini library. Dabble is used as the controller. The code can be tweaked